reduce jest copy-paste setup
