+++ packages/react-app-rewired/scripts/test.js
@@ -12,36 +12,42 @@ process.on("unhandledRejection", err => {
   throw err;
 });
 
+const jest = require("jest");
+const path = require("path");
 const paths = require("./utils/paths");
 // Ensure environment variables are read.
 require(paths.scriptVersion + "/config/env");
 
-const jest = require("jest");
 const argv = process.argv.slice(2);
-
 // Watch unless on CI or in coverage mode
 if (!process.env.CI && argv.indexOf("--coverage") < 0) {
   argv.push("--watch");
 }
 
-
-const createJestConfig = require("./utils/createJestConfig");
+const createJestConfig = require(paths.scriptVersion + "/scripts/utils/createJestConfig");
+const rewireJestConfig = require("./utils/rewireJestConfig");
 const override = require(paths.configOverrides);
 const overrideFn = (typeof override === 'function' || typeof override.jest !== 'function')
   ? (config) => config
   : override.jest;
 
-const path = require("path");
+// hide overrides in package.json for CRA's original createJestConfig
+const packageJson = require(paths.appPackageJson);
+const jestOverrides = packageJson.jest;
+delete packageJson.jest;
+
+const config = createJestConfig(
+  relativePath => path.resolve(paths.appPath, "node_modules", paths.scriptVersion, relativePath),
+  path.resolve(paths.appSrc, ".."),
+  false
+);
+
+// restore overrides for rewireJestConfig
+packageJson.jest = jestOverrides;
 
 argv.push(
   "--config",
-  JSON.stringify(
-    overrideFn(createJestConfig(
-      relativePath => path.resolve(__dirname, "..", relativePath),
-      path.resolve(paths.appSrc, ".."),
-      false
-    ))
-  )
+  JSON.stringify(overrideFn(rewireJestConfig(config)))
 );
 
 jest.run(argv);
