+++ packages/react-app-rewired/scripts/utils/babelTransform.js
@@ -1,11 +1,4 @@
-/**
- * Copyright (c) 2014-present, Facebook, Inc. All rights reserved.
- *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree. An additional grant
- * of patent rights can be found in the PATENTS file in the same directory.
- */
-
+// TODO: remove file in major release
 const babelJest = require('babel-jest');
 
 const customPlugins = [];
