--- packages/react-app-rewired/scripts/utils/createJestConfig.js
@@ -1,71 +1,0 @@
-// @remove-file-on-eject
-/**
- * Copyright (c) 2015-present, Facebook, Inc.
- * All rights reserved.
- *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree. An additional grant
- * of patent rights can be found in the PATENTS file in the same directory.
- */
-'use strict';
-
-const fs = require('fs');
-const chalk = require('chalk');
-const paths = require('./paths');
-
-module.exports = (resolve, rootDir, isEjecting) => {
-  // Use this instead of `paths.testsSetup` to avoid putting
-  // an absolute filename into configuration after ejecting.
-  const setupTestsFile = fs.existsSync(paths.testsSetup)
-    ? '<rootDir>/src/setupTests.js'
-    : undefined;
-
-  // TODO: I don't know if it's safe or not to just use / as path separator
-  // in Jest configs. We need help from somebody with Windows to determine this.
-  const config = {
-    collectCoverageFrom: ['src/**/*.{js,jsx}'],
-    setupFiles: [require.resolve(paths.scriptVersion + '/config/polyfills.js')],
-    setupTestFrameworkScriptFile: setupTestsFile,
-    testMatch: [
-      '<rootDir>/src/**/__tests__/**/*.js?(x)',
-      '<rootDir>/src/**/?(*.)(spec|test).js?(x)',
-    ],
-    testEnvironment: 'node',
-    testURL: 'http://localhost',
-    transform: {
-      '^.+\\.(js|jsx)$': isEjecting
-        ? '<rootDir>/node_modules/babel-jest'
-        : resolve('scripts/utils/babelTransform.js'),
-      '^.+\\.css$': require.resolve(paths.scriptVersion + '/config/jest/cssTransform.js'),
-      '^(?!.*\\.(js|jsx|css|json)$)': require.resolve(paths.scriptVersion + '/config/jest/fileTransform.js'),
-    },
-    transformIgnorePatterns: ['[/\\\\]node_modules[/\\\\].+\\.(js|jsx)$'],
-    moduleNameMapper: {
-      '^react-native$': 'react-native-web',
-    },
-    moduleFileExtensions: ['web.js', 'js', 'json', 'web.jsx', 'jsx', 'node'],
-  };
-  if (rootDir) {
-    config.rootDir = rootDir;
-  }
-  const overrides = Object.assign({}, require(paths.appPackageJson).jest);
-  
-  // Jest configuration in package.json will be added to the the default config
-  Object.keys(overrides)
-    .forEach(key => {
-      //We don't overwrite the default config, but add to each property if not a string
-      if(config[key]) {
-        if(typeof overrides[key] === 'string') {
-          config[key] = overrides[key];
-        } else if(Array.isArray(overrides[key])) {
-          config[key] = overrides[key].concat(config[key]);
-        }
-        else if(typeof overrides[key] === 'object') {
-          config[key] = Object.assign({}, overrides[key], config[key]);
-        }
-      } else {
-        config[key] = overrides[key];
-      }
-    });
-  return config;
-};
